refactor: fix typos, magic numbers, and props
