+++ src/components/Game/gameLogic/WinComponent.tsx
@@ -1,20 +1,26 @@
 import { Link } from 'react-router-dom';
-import { GamePopup } from '../GamePopup';
+import { GamePopup } from '../modal/GamePopup';
 import { NonogramRaw } from '../gameUtils/types';
 import { WinContent } from './WinContent';
 
+const WinModalCaptions = {
+    modalTitle: "Hooray, you've solved the nonogram!",
+    modalDismissChoise: 'To Catalog',
+    modalAcceptChoice: 'Try Again',
+} as const;
 export function WinComponent({ nonogramRaw }: { nonogramRaw: NonogramRaw }): JSX.Element {
-    const WIN_MODAL_CAPTIONS = {
-        modalTitle: "Hooray, you've solved the nonogram!",
-        modalDismissChoise: 'To Catalog',
-        modalAcceptChoice: 'Try Again',
-    };
     return (
-        <GamePopup captions={WIN_MODAL_CAPTIONS}>
-            <WinContent nonogramRaw={nonogramRaw} />
-            <Link to="/catalog" className="btn btn-outline-dark">
-                {WIN_MODAL_CAPTIONS.modalDismissChoise}
-            </Link>
+        <GamePopup captions={WinModalCaptions}>
+            <>
+                <div className="modal-body">
+                    <WinContent nonogramRaw={nonogramRaw} />
+                </div>
+                <div className="modal-footer">
+                    <Link to="/catalog" className="btn btn-outline-dark">
+                        {WinModalCaptions.modalDismissChoise}
+                    </Link>
+                </div>
+            </>
         </GamePopup>
     );
 }
