+++ src/components/Game/modal/GamePopup.tsx
@@ -5,17 +5,21 @@ interface ModalCaptions {
     modalDismissChoise: string;
     modalAcceptChoice: string;
 }
+const modalStyle = {
+    display: 'block',
+    backgroundColor: 'rgba(0,0,0,0.5)',
+};
 
 export function GamePopup({
     captions,
     children,
 }: {
     captions: ModalCaptions;
-    children: JSX.Element[];
+    children: JSX.Element;
 }): JSX.Element {
     const [isVisible, setIsVisible] = useState(true);
     const modalHandler = () => {
-        setIsVisible((prevoius) => !prevoius);
+        setIsVisible((previous) => !previous);
     };
     return (
         <>
@@ -24,10 +28,7 @@ export function GamePopup({
                 <div
                     className="modal fade show"
                     tabIndex={-1}
-                    style={{
-                        display: 'block',
-                        backgroundColor: 'rgba(130,130,130,0.5)',
-                    }}
+                    style={modalStyle}
                     onClick={modalHandler}
                     role="presentation"
                 >
@@ -48,8 +49,7 @@ export function GamePopup({
                                     aria-label="Close"
                                 />
                             </div>
-                            <div className="modal-body">{children[0]}</div>
-                            <div className="modal-footer">{children[1]}</div>
+                            {children}
                         </div>
                     </div>
                 </div>
